--- .gitignore
@@ -1,1 +1,0 @@
-GPIO_interface.h
