--- GPIO_Config.h
@@ -1,13 +1,0 @@
-////////////////////////////////////////////////////
-///////      Ain Shams University         /////////
-//////     Intro To Embedded Project     /////////
-/////           Layer:  MCAL            /////////
-////                GPIO               /////////
-///         DATE:   16-3-2025         /////////
-//         AUTHOR: Hussein Bahaa     /////////
-/////////////////////////////////////////////   
-
-#ifndef GPIO_CONFIG_H
-#define GPIO_CONFIG_H
-
-#endif /*GPIO_CONFIG_H*/
